backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-dotenv==1.0.0
httpx>=0.23.0,<0.24.0  # Compatible with supabase
pydantic==1.10.7
email-validator==2.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
supabase==1.0.3
sqlalchemy==2.0.12
asyncpg==0.27.0  # Using asyncpg instead of psycopg2-binary
alembic==1.10.4
pytest==7.3.1
pytest-asyncio==0.21.0
